+++ src/structlog/processors.py
@@ -731,6 +731,9 @@ class CallsiteParameter(enum.Enum):
 
     .. versionadded:: 25.5.0
        `QUAL_NAME` parameter.
+
+    .. versionadded:: 26.1.0
+       `QUAL_MODULE` parameter.
     """
 
     #: The full path to the python source file of the callsite.
@@ -742,6 +745,8 @@ class CallsiteParameter(enum.Enum):
     #: of `logging.LogRecord` objects and will be the basename, without
     #: extension, of the full path to the python source file of the callsite.
     MODULE = "module"
+    #: The fully qualified import name of the module of the callsite.
+    QUAL_MODULE = "qual_module"
     #: The name of the function that the callsite was in.
     FUNC_NAME = "func_name"
     #: The qualified name of the callsite (includes scope and class names).
@@ -777,6 +782,10 @@ def _get_callsite_func_name(module: str, frame: FrameType) -> Any:
 
 def _get_callsite_qual_name(module: str, frame: FrameType) -> Any:
     return frame.f_code.co_qualname  # will crash on Python <3.11
+
+
+def _get_callsite_qual_module(module: str, frame: FrameType) -> Any:
+    return frame.f_globals.get("__name__", "__main__")
 
 
 def _get_callsite_lineno(module: str, frame: FrameType) -> Any:
@@ -854,6 +863,7 @@ class CallsiteParameterAdder:
         CallsiteParameter.PATHNAME: _get_callsite_pathname,
         CallsiteParameter.FILENAME: _get_callsite_filename,
         CallsiteParameter.MODULE: _get_callsite_module,
+        CallsiteParameter.QUAL_MODULE: _get_callsite_qual_module,
         CallsiteParameter.FUNC_NAME: _get_callsite_func_name,
         CallsiteParameter.QUAL_NAME: _get_callsite_qual_name,
         CallsiteParameter.LINENO: _get_callsite_lineno,